// src/helux/src/typing.d.ts
export type PrimitiveItem = number | string;

export type PrimitiveSymItem = PrimitiveItem | symbol;

export type Dict<T extends any = any> = Record<PrimitiveSymItem, T>;

export type DictN<T extends any = any> = Record<number, T>;

export type Fn<T extends any = any> = (...args: any[]) => T;

export type SharedObject<T extends Dict = any> = T;

export type EenableReactive = boolean;

/** 是否在执行计算中，如是同步的计算结果，hook里此值总是 false，不会产生变化 */
export type IsComputing = boolean;
// ...
export interface IFnParams {
  isFirstCall: boolean;
}

export interface IAsyncTaskParams<S extends any = any> extends IFnParams {
  source: S;
}

export type ComputedResult<T extends Dict = Dict> = T;

export type ComputedFn<T extends Dict = Dict> = (params: IFnParams) => T;

export interface IUnmountInfo {
  t: number;
  /** mount count, 第一次挂载或第二次挂载 */
  c: 1 | 2;
  /**
   * @deprecated
   * 前一个实例 id，已无意义，后续会移除
   */
  prev: number;
}

export type FnType = 'watch' | 'computed';

export type ScopeType = 'static' | 'hook';

export type AsyncType = 'source' | 'task' | 'normal';

export type ReanderStatus = '1' | '2';

export type MountStatus = 1 | 2 | 3;

export interface IFnCtx {
  /** 
   * 计算函数本体，即透传给 derive 的回调函数
   */
  fn: Fn;
  /**
   * 函数唯一标记 key
   */
  fnKey: string;
  /**
   *  deriveAsync/useDeriveAsync 传入的第一个回调函数
   */
  sourceFn: Fn;
  /**
   * default: true，是否是处于第一层的函数，使用了其他计算结果时就会表标记为 false
   */
  isFirstLevel: boolean;
  isComputing: boolean;
  remainRunCount: number;
  careComputeStatus: boolean;
  /**
   * default: false ，是否对计算结果开启记录读依赖功能，此功能仅针对 hook 里使用 useComputed 有效
   */
  enableRecordResultDep: boolean;
  /**
   * 直接依赖此函数的下一层函数列表，如其他函数使用了此函数的返回结果（包括中转返回结果），则它们的 key 会被记录到这里
   */
  nextLevelFnKeys: string[];
  /** 此函数依赖的上一层函数列表，如此函数内部使用了其他函数的返回结果，则把其他函数的 key 会被记录到这里 */
  prevLevelFnKeys: string[];
  /** 未挂载 已挂载 已卸载 */
  mountStatus: MountStatus;
  depKeys: string[];
  /**
   * 计算函数返回的原始结果，总是指向第一次计算返回的结果
   */
  result: Dict;
  /**
   * 提供给 hook 函数读取的代理结果
   */
  proxyResult: Dict;
  fnType: FnType;
  scopeType: ScopeType;
  /** work for hook computed fnCtx */
  updater: Fn;
  /** work for hook computed fnCtx */
  isResultReaded: boolean;
  /** 只要结果曾经读取过就记录为 true */
  isResultReadedOnce: boolean;
  /**
   * 是否返回了上游的计算结算，方便为计算结果中转机制服务
   * work for computed result transfer mechanism
   */
  returnUpstreamResult: boolean;
  /** work for hook computed fnCtx */
  renderStatus: ReanderStatus;
  /** fn ctx created timestamp */
  createTime: number;
  /** work for hook computed fnCtx  */
  shouldReplaceResult: boolean;
  /**
   * 是否是异步的计算函数，使用了异步计算结果、返回了异步计算结果、返回了 asyncTask，满足任意一种情况都会标记为 true
   */
  isAsync: boolean;
  /** 是否是一个中转结果的异步函数，内部用的标记 */
  isAsyncTransfer: boolean;
  asyncType: AsyncType,
}